+++ spec/spec_helper.rb
@@ -2,6 +2,8 @@ require 'sms_box'
 require 'equivalent-xml'
 require 'nokogiri'
 
+Dir[File.expand_path(File.join(File.dirname(__FILE__),'support','**','*.rb'))].each {|f| require f}
+
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
   config.treat_symbols_as_metadata_keys_with_true_values = true