+++ spec/unit/sms_box/client_spec.rb
@@ -1,16 +1,5 @@
 require 'spec_helper'
 
-WEBSEND_RESPONSE = '<?xml version="1.0" encoding="UTF-8" ?>
-<SMSBoxXMLReply>
-  <ok/>
-  <command name="WEBSEND">
-    <receiver status="ok">+41790000001</receiver>
-    <receiver status="ok">+41790000002</receiver>
-    <receiver status="ok">+41790000003</receiver>
-  </command>
-  <requestUid>xml9677322</requestUid>
-</SMSBoxXMLReply>'
-
 module SMSBox
   class DummyResponse < XMLResponse; end